// apps/mobile/src/shared/types.ts
export type PostCategory = 'نقاش' | 'سؤال' | 'حدث' | 'استطلاع رأي';

export interface Comment {
  id: number;
  userId: number;
  username: string;
  avatar: string;
  content: string;
  date: string;
}

export interface PollOption {
  option: string;
  votes: number[];
}

export interface Post {
  id: number;
  circleId: number;
  userId: number;
  username: string;
  avatar: string;
  title?: string;
  content: string;
  category: PostCategory;
  date: string;
  likes: number[];
  comments: Comment[];
  isPinned?: boolean;
  pollOptions?: PollOption[];
}